# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Event:
    properties:
      description:
        type: string
      eventEnd:
        type: string
      eventMaxCapacity:
        type: integer
      eventStart:
        type: string
      eventroId:
        type: string
      fee:
        type: string
      food:
        type: string
      id:
        type: integer
      imageUrl:
        type: string
      location:
        type: string
      name:
        type: string
      registrationEnd:
        type: string
      registrationRequired:
        type: boolean
      show:
        type: boolean
      signupLink:
        type: string
    type: object
...
  models.Profile:
    properties:
      email:
        type: string
      eventro_key:
        type: string
      id:
        type: integer
      linkedin:
        type: string
      name:
        type: string
      photo:
        type: string
      rank:
        type: string
      team:
        $ref: '#/definitions/models.Team'
      team_id:
        type: integer
      title:
        type: string
    type: object
...
  models.Team:
    properties:
      id:
        type: integer
      team_name:
        type: string
    type: object
...
info:
  contact:
    name: THS Armada Development
    url: https://github.com/armada-ths
  description: REST API powering the THS Armada career fair website (armada.nu) and
    its admin panel.
  title: ArmadaCMS API
  version: "1.0"
paths:
...
  /eventromembers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Profile'
            type: array
        "500":
          description: Failed to fetch from Eventro
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Sync & list member profiles from Eventro
      tags:
      - eventro
...
  /events:
    get:
      parameters:
      - description: Pagination range, e.g. [0,24]
        in: query
        name: range
        type: string
      - description: Sort, e.g. [\
        in: query
        name: sort
        type: string
      - description: Filter, e.g. {\
        in: query
        name: filter
        type: string
      - description: If true, only return events with show=true
        in: query
        name: public
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Content-Range:
              description: events 0-24/100
              type: string
          schema:
            items:
              $ref: '#/definitions/models.Event'
            type: array
      summary: List events
      tags:
      - events
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Event name
        in: formData
        name: name
        required: true
        type: string
      - description: Location
        in: formData
        name: location
        required: true
        type: string
      - description: Start time (RFC3339)
        in: formData
        name: eventStart
        required: true
        type: string
      - description: End time (RFC3339)
        in: formData
        name: eventEnd
        required: true
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      - description: Food information
        in: formData
        name: food
        type: string
      - description: Registration required
        in: formData
        name: registrationRequired
        type: boolean
      - description: Fee information
        in: formData
        name: fee
        type: string
      - description: Signup link URL
        in: formData
        name: signupLink
        type: string
      - description: Max capacity
        in: formData
        name: eventMaxCapacity
        type: integer
      - description: Registration end time (RFC3339)
        in: formData
        name: registrationEnd
        type: string
      - description: Whether to show the event publicly
        in: formData
        name: show
        type: boolean
      - description: Event image
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Event'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Create failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create event
      tags:
      - events
  /events/{id}:
    delete:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted
          schema:
            type: string
        "404":
          description: Event not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete event
      tags:
      - events
    get:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Event'
        "404":
          description: Event not found
          schema:
            type: string
      summary: Get event by ID
      tags:
      - events
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Event name
        in: formData
        name: name
        type: string
      - description: Location
        in: formData
        name: location
        type: string
      - description: Start time (RFC3339)
        in: formData
        name: eventStart
        type: string
      - description: End time (RFC3339)
        in: formData
        name: eventEnd
        type: string
      - description: Whether to show the event publicly
        in: formData
        name: show
        type: boolean
      - description: Event image
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Event'
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
        "500":
          description: Update failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update event
      tags:
      - events
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer JWT token obtained from POST /api/v1/login. Format: "Bearer
      <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"